requirements.txt
flasgger==0.9.7.1
Flask==3.0.3
PyJWT==2.9.0
PyMySQL==1.1.0
PyYAML==6.0.2
Requests==2.32.3
